Move functionality in exec.go to bundle.go
